+++ libs/chain-signatures/contract/src/primitives/participants.rs
